Remove Database::get_top_clusters_by_score
It always returned an empty set and provided no value.
