+++ src/csv_indicator.rs
@@ -4,7 +4,7 @@ use diesel::{ExpressionMethods, OptionalExtension, QueryDsl};
 use diesel_async::{pg::AsyncPgConnection, RunQueryDsl};
 
 use crate::{
-    schema::{csv_column_list, csv_indicator, csv_whitelist},
+    schema::{csv_column_list, csv_whitelist},
     Error,
 };
 
@@ -55,59 +55,5 @@ pub(crate) async fn get_whitelists(
     Ok(whitelists
         .into_iter()
         .map(|(name, value)| (*whitelist_names_indices.get(&name).expect(""), value))
-        .collect())
-}
-
-pub(crate) async fn get_csv_indicators(
-    conn: &mut AsyncPgConnection,
-    model_id: i32,
-    column_indices: &[usize],
-) -> Result<HashMap<usize, String>, Error> {
-    use csv_column_list::dsl as list_d;
-    use csv_indicator::dsl as i_d;
-
-    let Some(Some(indicator_names)) = list_d::csv_column_list
-        .select(list_d::column_indicator)
-        .filter(list_d::model_id.eq(model_id))
-        .get_result::<Option<Vec<Option<String>>>>(conn)
-        .await
-        .optional()?
-    else {
-        return Ok(HashMap::new());
-    };
-    let indicator_names_indices: HashMap<String, usize> = column_indices
-        .iter()
-        .filter_map(|i| {
-            indicator_names.get(*i).and_then(|n| {
-                n.as_ref().and_then(|n| {
-                    if n.is_empty() {
-                        None
-                    } else {
-                        Some((n.clone(), *i))
-                    }
-                })
-            })
-        })
-        .collect();
-    let indicator_names: Vec<String> = indicator_names_indices
-        .keys()
-        .map(std::clone::Clone::clone)
-        .collect();
-    let indicators = i_d::csv_indicator
-        .select((i_d::name, i_d::list))
-        .filter(i_d::name.eq_any(&indicator_names))
-        .get_results::<(String, String)>(conn)
-        .await
-        .optional()?
-        .unwrap_or_default();
-
-    Ok(indicators
-        .into_iter()
-        .map(|(name, value)| {
-            (
-                *indicator_names_indices.get(&name).expect("already checked"),
-                value,
-            )
-        })
         .collect())
 }
